chore(index.html, /js/*): reorganized business logic and imports
